requirement.txt
google.generativeai
python-dotenv
streamlit
seaborn
plotly
